move zarr v3 code to root
