+++ src/zarr/attributes.py
@@ -3,8 +3,8 @@ from collections.abc import MutableMapping
 from typing import TYPE_CHECKING, Any, Union
 
 if TYPE_CHECKING:
-    from zarr.v3.group import Group
-    from zarr.v3.array import Array
+    from zarr.group import Group
+    from zarr.array import Array
 
 
 class Attributes(MutableMapping[str, Any]):
