+++ src/zarr/chunk_grids.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 from typing import TYPE_CHECKING, Any, Dict
 from dataclasses import dataclass
-from zarr.v3.abc.metadata import Metadata
+from zarr.abc.metadata import Metadata
 
-from zarr.v3.common import (
+from zarr.common import (
     JSON,
     ChunkCoords,
     ChunkCoordsLike,
