+++ src/zarr/chunk_key_encodings.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 from abc import abstractmethod
 from typing import TYPE_CHECKING, Dict, Literal, cast
 from dataclasses import dataclass
-from zarr.v3.abc.metadata import Metadata
+from zarr.abc.metadata import Metadata
 
-from zarr.v3.common import (
+from zarr.common import (
     JSON,
     ChunkCoords,
     parse_named_configuration,
