+++ src/zarr/codecs/blosc.py
@@ -9,15 +9,15 @@ import numcodecs
 import numpy as np
 from numcodecs.blosc import Blosc
 
-from zarr.v3.abc.codec import BytesBytesCodec
-from zarr.v3.codecs.registry import register_codec
-from zarr.v3.common import parse_enum, parse_named_configuration, to_thread
+from zarr.abc.codec import BytesBytesCodec
+from zarr.codecs.registry import register_codec
+from zarr.common import parse_enum, parse_named_configuration, to_thread
 
 if TYPE_CHECKING:
     from typing import Dict, Optional
     from typing_extensions import Self
-    from zarr.v3.common import JSON, ArraySpec, BytesLike
-    from zarr.v3.config import RuntimeConfiguration
+    from zarr.common import JSON, ArraySpec, BytesLike
+    from zarr.config import RuntimeConfiguration
 
 
 class BloscShuffle(Enum):
