+++ src/zarr/codecs/gzip.py
@@ -4,15 +4,15 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 from numcodecs.gzip import GZip
-from zarr.v3.abc.codec import BytesBytesCodec
-from zarr.v3.codecs.registry import register_codec
-from zarr.v3.common import parse_named_configuration, to_thread
+from zarr.abc.codec import BytesBytesCodec
+from zarr.codecs.registry import register_codec
+from zarr.common import parse_named_configuration, to_thread
 
 if TYPE_CHECKING:
     from typing import Optional, Dict
     from typing_extensions import Self
-    from zarr.v3.common import JSON, ArraySpec, BytesLike
-    from zarr.v3.config import RuntimeConfiguration
+    from zarr.common import JSON, ArraySpec, BytesLike
+    from zarr.config import RuntimeConfiguration
 
 
 def parse_gzip_level(data: JSON) -> int:
