+++ src/zarr/indexing.py
@@ -4,7 +4,7 @@ import itertools
 import math
 from typing import Iterator, List, NamedTuple, Optional, Tuple
 
-from zarr.v3.common import ChunkCoords, Selection, SliceSelection, product
+from zarr.common import ChunkCoords, Selection, SliceSelection, product
 
 
 def _ensure_tuple(v: Selection) -> SliceSelection:
