--- src/zarr/v3/codecs/__init__.py
@@ -1,9 +1,0 @@
-from __future__ import annotations
-
-from zarr.v3.codecs.blosc import BloscCodec, BloscCname, BloscShuffle  # noqa: F401
-from zarr.v3.codecs.bytes import BytesCodec, Endian  # noqa: F401
-from zarr.v3.codecs.crc32c_ import Crc32cCodec  # noqa: F401
-from zarr.v3.codecs.gzip import GzipCodec  # noqa: F401
-from zarr.v3.codecs.sharding import ShardingCodec, ShardingCodecIndexLocation  # noqa: F401
-from zarr.v3.codecs.transpose import TransposeCodec  # noqa: F401
-from zarr.v3.codecs.zstd import ZstdCodec  # noqa: F401
