--- src/zarr/v3/store/__init__.py
@@ -1,5 +1,0 @@
-# flake8: noqa
-from zarr.v3.store.core import StorePath, StoreLike, make_store_path
-from zarr.v3.store.remote import RemoteStore
-from zarr.v3.store.local import LocalStore
-from zarr.v3.store.memory import MemoryStore
